Revert "enh: optimize cache last and other functions performance"
