+++ include/libs/nodes/plannodes.h
@@ -40,13 +40,6 @@ typedef enum EGroupAction {
   GROUP_ACTION_CLEAR
 } EGroupAction;
 
-typedef enum EMergeType {
-  MERGE_TYPE_SORT = 1,
-  MERGE_TYPE_NON_SORT,
-  MERGE_TYPE_COLUMNS,
-  MERGE_TYPE_MAX_VALUE
-} EMergeType;
-
 typedef struct SLogicNode {
   ENodeType          type;
   bool               dynamicOp;
@@ -145,7 +138,6 @@ typedef struct SAggLogicNode {
   bool       hasGroupKeyOptimized;
   bool       isGroupTb;
   bool       isPartTb;  // true if partition keys has tbname
-  bool       hasGroup;
 } SAggLogicNode;
 
 typedef struct SProjectLogicNode {
@@ -229,8 +221,6 @@ typedef struct SMergeLogicNode {
   SNodeList* pInputs;
   int32_t    numOfChannels;
   int32_t    srcGroupId;
-  bool       colsMerge;
-  bool       needSort;
   bool       groupSort;
   bool       ignoreGroupId;
   bool       inputWithGroupId;
@@ -542,7 +532,6 @@ typedef struct SExchangePhysiNode {
 
 typedef struct SMergePhysiNode {
   SPhysiNode node;
-  EMergeType type;
   SNodeList* pMergeKeys;
   SNodeList* pTargets;
   int32_t    numOfChannels;
