+++ source/libs/executor/src/mergejoinoperator.c
@@ -239,7 +239,7 @@ SOperatorInfo* createMergeJoinOperatorInfo(SOperatorInfo** pDownstream, int32_t 
     numOfDownstream = 2;
   } else {
     pInfo->downstreamResBlkId[0] = getOperatorResultBlockId(pDownstream[0], 0);
-    pInfo->downstreamResBlkId[1] = getOperatorResultBlockId(pDownstream[1], 0);
+    pInfo->downstreamResBlkId[1] = getOperatorResultBlockId(pDownstream[1], 1);
   }
 
   int32_t      numOfCols = 0;
