+++ source/libs/nodes/src/nodesMsgFuncs.c
@@ -2690,7 +2690,6 @@ enum {
   PHY_MERGE_CODE_GROUP_SORT,
   PHY_MERGE_CODE_IGNORE_GROUP_ID,
   PHY_MERGE_CODE_INPUT_WITH_GROUP_ID,
-  PHY_MERGE_CODE_TYPE,
 };
 
 static int32_t physiMergeNodeToMsg(const void* pObj, STlvEncoder* pEncoder) {
@@ -2717,9 +2716,6 @@ static int32_t physiMergeNodeToMsg(const void* pObj, STlvEncoder* pEncoder) {
   }
   if (TSDB_CODE_SUCCESS == code) {
     code = tlvEncodeBool(pEncoder, PHY_MERGE_CODE_INPUT_WITH_GROUP_ID, pNode->inputWithGroupId);
-  }
-  if (TSDB_CODE_SUCCESS == code) {
-    code = tlvEncodeI32(pEncoder, PHY_MERGE_CODE_TYPE, pNode->type);
   }
 
   return code;
@@ -2755,9 +2751,6 @@ static int32_t msgToPhysiMergeNode(STlvDecoder* pDecoder, void* pObj) {
         break;
       case PHY_MERGE_CODE_INPUT_WITH_GROUP_ID:
         code = tlvDecodeBool(pTlv, &pNode->inputWithGroupId);
-        break;
-      case PHY_MERGE_CODE_TYPE:
-        code = tlvDecodeI32(pTlv, (int32_t*)&pNode->type);
         break;
       default:
         break;
